+++ src/convert_roms/test_chd_parents_to_chd_should_not_touch_parent.rs
@@ -3,7 +3,7 @@ use super::super::import_dats;
 use super::super::import_roms;
 use super::*;
 use std::path::{Path, PathBuf};
-use std::{env, time::SystemTime};
+use std::time::SystemTime;
 use tempfile::{NamedTempFile, TempDir};
 use tokio::fs;
 
@@ -13,17 +13,6 @@ async fn test() {
     let _guard = MUTEX.lock().await;
 
     let test_directory = Path::new("tests");
-    // TODO: Audit that the environment access only happens in single-threaded code.
-    unsafe {
-        env::set_var(
-            "PATH",
-            format!(
-                "{}:{}",
-                test_directory.as_os_str().to_str().unwrap(),
-                env::var("PATH").unwrap()
-            ),
-        )
-    };
     let progress_bar = ProgressBar::hidden();
 
     let db_file = NamedTempFile::new().unwrap();